Completed part 2a exercise 2.5: separate module
